+++ src/test/java/roomescape/ReservationDAOTest.java
@@ -5,6 +5,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
+import roomescape.dao.ReservationDAO;
+import roomescape.entity.ReservationEntity;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
