+++ src/test/java/roomescape/ReservationEntityTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import roomescape.entity.ReservationEntity;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
